Split spring-graphql auto configuration
